add Windows-only font height fix
A given font height on Mac OS X looks three points larger on Windows. I had this same problem with Vim. This code overrides the setting for Mac OS X if it detects that we're on Windows. (I'm just guessing that "mingw" is the string to look for within the output of system-configuration, since "win" doesn't appear.)