disable copy to system clipboard on select
I sometimes copy something in another program, select the text I want to replace, and then paste. But before this change, that just pasted the same text back, because it was automatically copied when it was selected. That behavior was counterproductive. I'm fine with explicitly saying when I'm copying to the clipboard, like Vim does with "* . (Though I have to use "+ with Evil to access the clipboard; "* doesn't work for some reason.)

by disabling x-select-enable-clipboard

http://www.gnu.org/software/emacs/manual/html_node/emacs/Clipboard.html