Remove finished TODO
This TODO should have been moved earlier to the bottom of the plugin-setting section. But that doesn’t matter now.

I have done as the TODO asked. But there might still be room for improvement. I still want to look up if it is better to use `require` and `provide`, like https://github.com/juanjux/emacs-dotfiles does it, than `load`.