Disable mouse-wheel-progressive-speed
It makes scrolling skip more lines if more wheel events are given quickly. But that makes it harder to predict how far I will scroll, and tends to make most scrolls too large. The downside of disabling it is that scrolling all the way to the top or bottom is harder.

Sadly, this doesn’t make the scrolling any smoother – Emacs still has a too-high debounce delay for mouse scroll events.